Bug fixed

+++ app/src/main/java/com/example/androidproject/api/ApiInterface.kt
@@ -7,7 +7,7 @@ import retrofit2.http.*
 interface ApiInterface {
 
     @GET("/restaurants/{id}")
-    suspend fun getRestaurantById(@Path("id") id: Int): Response<Restaurant>
+    suspend fun getRestaurantById(@Path("id") id: Long): Response<Restaurant>
 
     @GET("/restaurants")
     suspend fun getAllRestaurants(@QueryMap params: Map<String, String>): Response<Restaurants>
